server/requirements.txt
asgiref==3.3.4
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==5.0.9
Django==3.2.3
django-cors-headers==3.7.0
django-extensions==3.1.1
django-rest-framework==0.1.0
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Faker==8.1.4
idna==2.10
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
Pygments==2.9.0
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
